Fix significant performance issue on WSL2
If we didn't find any snippets for a particular filetype, we'd search
through files again every time we expanded a snippet. Now, we record the
fact that you attempted the search so you don't have to do so again.

The reason this is an issue on WSL2 is that the glob and realpath
functions are extremely slow on a mounted drive.

You'll still hit the hitch the very first time expanding a snippet in a
session, but it should be real fast after that.